component ID, azure in separate dir
