+++ src/index.ts
@@ -53,4 +53,4 @@ class Agent {
 
 export default new Agent();
 export { default as ContextManager } from './trace/context/ContextManager';
-export { default as AzureHttpTriggerPlugin } from './plugins/AzureHttpTriggerPlugin';
+export { default as AzureHttpTriggerPlugin } from './azure/AzureHttpTriggerPlugin';
